test(iceoryx2): split Local tests out of integration suite, run IPC by default
The iceoryx2 `Local` variant is fully in-process — those round-trip tests
were unit tests living under a misleading filename. Move them to a new
`local_tests.rs` so they run with `cargo test --features iceoryx2-beta`,
not just the gated integration job. Drop the duplicate Local smoke tests
that had drifted into `mod.rs::tests`.

The two genuine cross-process IPC tests (`test_ipc_round_trip`,
`test_late_joiner_with_history`) now live alone in `integration_tests.rs`,
which is fully feature-gated at the file level. The `#[ignore]` attributes
are removed: the tests pass reliably on Linux runners with a writable
`/dev/shm`, and the workflow already path-triggers on adapter changes, so
hiding them behind a manual checkbox added no safety.

CI is collapsed accordingly — the manual `iceoryx2-ipc-ignored` job is
gone; the standard job now runs both Local and IPC tests in one cargo
invocation.
