+++ wingfoil/src/adapters/iceoryx2/mod.rs
@@ -268,18 +268,17 @@ mod write;
 pub use read::*;
 pub use write::*;
 
-#[cfg(any(test, feature = "iceoryx2-integration-test"))]
-pub mod integration_tests;
+#[cfg(test)]
+mod local_tests;
+
+#[cfg(test)]
+mod integration_tests;
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::nodes::{NodeOperators, StreamOperators};
-    use crate::{Burst, Graph, RunFor, RunMode, ticker};
+    use crate::{Burst, Graph, RunFor, RunMode};
     use iceoryx2::prelude::ZeroCopySend;
-    use std::sync::Arc;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-    use std::time::Duration;
 
     #[test]
     fn test_burst_creation() {
@@ -293,172 +292,6 @@ mod tests {
     #[derive(Debug, Clone, Copy, Default, ZeroCopySend)]
     struct TestData {
         value: u64,
-    }
-
-    #[test]
-    fn test_local_pubsub_smoke() {
-        static COUNTER: AtomicUsize = AtomicUsize::new(0);
-        let n = COUNTER.fetch_add(1, Ordering::Relaxed);
-        let service_name = format!("wingfoil/test/local/{}/{n}", std::process::id());
-
-        let sub = iceoryx2_sub_with::<TestData>(&service_name, Iceoryx2ServiceVariant::Local);
-        let collected = sub.collapse().collect();
-
-        let upstream = ticker(Duration::from_millis(2)).produce(|| {
-            let mut b: Burst<TestData> = Burst::default();
-            b.push(TestData { value: 1 });
-            b
-        });
-        let pub_node = iceoryx2_pub_with(upstream, &service_name, Iceoryx2ServiceVariant::Local);
-
-        Graph::new(
-            vec![pub_node, collected.clone().as_node()],
-            RunMode::RealTime,
-            RunFor::Duration(Duration::from_millis(100)),
-        )
-        .run()
-        .unwrap();
-
-        let values: Vec<TestData> = collected
-            .peek_value()
-            .into_iter()
-            .map(|item| item.value)
-            .collect();
-        assert!(!values.is_empty(), "expected at least one received sample");
-    }
-
-    #[test]
-    fn test_local_slice_pubsub_smoke() {
-        static COUNTER: AtomicUsize = AtomicUsize::new(0);
-        let n = COUNTER.fetch_add(1, Ordering::Relaxed);
-        let service_name = format!("wingfoil/test/local/slice/{}/{n}", std::process::id());
-
-        let sub = iceoryx2_sub_slice(&service_name);
-        let collected = sub.collapse().collect();
-
-        let step = Arc::new(AtomicUsize::new(0));
-        let upstream = ticker(Duration::from_millis(20)).produce(move || {
-            let mut b: Burst<Vec<u8>> = Burst::default();
-            let s = step.fetch_add(1, Ordering::Relaxed);
-            if s.is_multiple_of(2) {
-                b.push(vec![1, 2, 3]);
-            } else {
-                b.push(vec![4, 5, 6, 7]);
-            }
-            b
-        });
-        let pub_node = iceoryx2_pub_slice(upstream, &service_name);
-
-        // Give a moment for ports to connect
-        std::thread::sleep(Duration::from_millis(100));
-
-        Graph::new(
-            vec![pub_node, collected.clone().as_node()],
-            RunMode::RealTime,
-            RunFor::Duration(Duration::from_secs(1)),
-        )
-        .run()
-        .unwrap();
-
-        let values: Vec<Vec<u8>> = collected
-            .peek_value()
-            .into_iter()
-            .map(|item| item.value)
-            .collect();
-        assert!(values.len() >= 2);
-        assert!(values.iter().any(|v| *v == vec![1, 2, 3]));
-        assert!(values.iter().any(|v| *v == vec![4, 5, 6, 7]));
-    }
-
-    #[test]
-    fn test_local_slice_signaled_pubsub_smoke() {
-        static COUNTER: AtomicUsize = AtomicUsize::new(0);
-        let n = COUNTER.fetch_add(1, Ordering::Relaxed);
-        let service_name = format!(
-            "wingfoil/test/local/slice/signaled/{}/{n}",
-            std::process::id()
-        );
-
-        let opts = Iceoryx2SubOpts {
-            variant: Iceoryx2ServiceVariant::Local,
-            mode: Iceoryx2Mode::Signaled,
-            ..Default::default()
-        };
-        let sub = iceoryx2_sub_slice_opts(&service_name, opts);
-        let collected = sub.collapse().collect();
-
-        let step = Arc::new(AtomicUsize::new(0));
-        let upstream = ticker(Duration::from_millis(20)).produce(move || {
-            let mut b: Burst<Vec<u8>> = Burst::default();
-            let s = step.fetch_add(1, Ordering::Relaxed);
-            if s.is_multiple_of(2) {
-                b.push(vec![10, 20]);
-            } else {
-                b.push(vec![30, 40, 50]);
-            }
-            b
-        });
-        let pub_node =
-            iceoryx2_pub_slice_with(upstream, &service_name, Iceoryx2ServiceVariant::Local);
-
-        std::thread::sleep(Duration::from_millis(100));
-
-        Graph::new(
-            vec![pub_node, collected.clone().as_node()],
-            RunMode::RealTime,
-            RunFor::Duration(Duration::from_secs(1)),
-        )
-        .run()
-        .unwrap();
-
-        let values: Vec<Vec<u8>> = collected
-            .peek_value()
-            .into_iter()
-            .map(|item| item.value)
-            .collect();
-        assert!(values.len() >= 2);
-        assert!(values.iter().any(|v| *v == vec![10, 20]));
-        assert!(values.iter().any(|v| *v == vec![30, 40, 50]));
-    }
-
-    #[test]
-    fn test_local_signaled_pubsub_smoke() {
-        static COUNTER: AtomicUsize = AtomicUsize::new(0);
-        let n = COUNTER.fetch_add(1, Ordering::Relaxed);
-        let service_name = format!("wingfoil/test/local/signaled/{}/{n}", std::process::id());
-
-        let opts = Iceoryx2SubOpts {
-            variant: Iceoryx2ServiceVariant::Local,
-            mode: Iceoryx2Mode::Signaled,
-            ..Default::default()
-        };
-        let sub = iceoryx2_sub_opts::<TestData>(&service_name, opts);
-        let collected = sub.collapse().collect();
-
-        let upstream = ticker(Duration::from_millis(20)).produce(|| {
-            let mut b: Burst<TestData> = Burst::default();
-            b.push(TestData { value: 123 });
-            b
-        });
-        let pub_node = iceoryx2_pub_with(upstream, &service_name, Iceoryx2ServiceVariant::Local);
-
-        std::thread::sleep(Duration::from_millis(100));
-
-        Graph::new(
-            vec![pub_node, collected.clone().as_node()],
-            RunMode::RealTime,
-            RunFor::Duration(Duration::from_secs(1)),
-        )
-        .run()
-        .unwrap();
-
-        let values: Vec<TestData> = collected
-            .peek_value()
-            .into_iter()
-            .map(|item| item.value)
-            .collect();
-        assert!(!values.is_empty(), "expected samples in signaled mode");
-        assert!(values.iter().all(|v| v.value == 123));
     }
 
     #[test]
